apps/backend/requirements.txt
playwright>=1.40.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0

# Additional recommended dependencies for production
aiofiles>=23.0.0  # For async file operations
pydantic>=2.0.0   # For data validation and settings management
structlog>=23.0.0 # For structured logging
tenacity>=8.0.0   # For retry logic
pytest>=7.0.0     # For testing
pytest-asyncio>=0.21.0  # For async testing
black>=23.0.0     # For code formatting
flake8>=6.0.0     # For code linting
mypy>=1.0.0       # For type checking
